+++ CHANGELOG.md
@@ -11,8 +11,10 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ### Added
 
+- Ability to print config for a file. [#158]
 - Support for `package.json` configuration file. [#245]
 
+[#158]: https://github.com/AdguardTeam/AGLint/issues/158
 [#245]: https://github.com/AdguardTeam/AGLint/issues/245
 
 ## [4.0.0-alpha.5] - 2025-11-03